requirements.txt
mysql-connector-python==8.0.30
requests==2.21.0
slack-bolt==1.7.0
beautifulsoup4==4.9.1
